+++ config/global.js
@@ -1,3 +1,5 @@
 module.exports = {
   'apiUrl':'http://localhost:3000',
+  'organizationBaseUrl' : 'org.local',
+  'organizationEnabled' : false
 };